implemented legion ECS

+++ src/map_builder.rs
@@ -58,7 +58,7 @@ impl MapBuilder {
 		use std::cmp::{min, max};
 		for y in min(y1, y2)..= max(y1, y2) {
 			if let Some(index) = self.map.try_index(Point::new(x, y)) {
-				self.map.tiles[index as usize] = TileType::Floor;
+				self.map.tiles[index] = TileType::Floor;
 			}
 		}
 	}
@@ -67,7 +67,7 @@ impl MapBuilder {
 		use std::cmp::{min, max};
 		for x in min(x1, x2)..= max(x1, x2) {
 			if let Some(index) = self.map.try_index(Point::new(x, y)) {
-				self.map.tiles[index as usize] = TileType::Floor;
+				self.map.tiles[index] = TileType::Floor;
 			}
 		}
 	}
